Gregor Gorjanc suggest to include zero-inflated binomial models in zeroinfl(). Still to do: gradients and hurdle() version.
Gregor GORJANC <gregor.gorjanc@bf.uni-lj.si>

I needed the ZIB model and I tried the package COZIGAM, but that
package works only with "smooth regressors" and can not be used for
other kind of effects, say ordinary regression or factors. I tried to
extend it, but it was to cumbersome. Therefore, I extended the
zeroinfl function from pscl to support ZIB model. I have not done
everything (gradient fucntion is missing, because I do not understand
it fully; some bits are to be checked, ...), but the code seems to
give sensible results. Would you find my additions interesting for
further work on this model?



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/countreg@12 edb9625f-4e0d-4859-8d74-9fd3b1da38cb